+++ packages/admin-api-client-data-provider/src/config.js
@@ -1,11 +1,17 @@
 import { providers } from "@data-provider/core";
-import { BASE_PATH, DEFAULT_PORT, DEFAULT_CLIENT_HOST } from "@mocks-server/admin-api-paths";
+import {
+  BASE_PATH,
+  DEFAULT_PORT,
+  DEFAULT_CLIENT_HOST,
+  DEFAULT_PROTOCOL,
+} from "@mocks-server/admin-api-paths";
 
 import TAG from "./tag";
 
 const DEFAULT_OPTIONS = {
   port: DEFAULT_PORT,
   host: DEFAULT_CLIENT_HOST,
+  protocol: DEFAULT_PROTOCOL,
 };
 
 export const configClient = (options) => {
@@ -15,7 +21,7 @@ export const configClient = (options) => {
   };
 
   providers.getByTag(TAG).config({
-    baseUrl: `http://${finalOptions.host}:${finalOptions.port}${BASE_PATH}`,
+    baseUrl: `${finalOptions.protocol}://${finalOptions.host}:${finalOptions.port}${BASE_PATH}`,
   });
 };