+++ packages/core/src/common/helpers.js
@@ -8,6 +8,9 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
+const path = require("path");
+const fsExtra = require("fs-extra");
+
 function arrayMerge(_destinationArray, sourceArray) {
   return sourceArray;
 }
@@ -36,9 +39,14 @@ function resolveWhenConditionPass(condition, { interval = 200, timeout = 2000 } 
   });
 }
 
+function readFileSync(filePath) {
+  return fsExtra.readFileSync(path.resolve(process.cwd(), filePath), "utf-8");
+}
+
 module.exports = {
   arrayMerge,
   deprecatedMessage,
   docsUrl,
   resolveWhenConditionPass,
+  readFileSync,
 };