+++ packages/plugin-admin-api/test/config-api.spec.js
@@ -67,6 +67,9 @@ describe("settings api", () => {
             path: "/admin",
             host: "0.0.0.0",
             port: 3110,
+            https: {
+              enabled: false,
+            },
           },
         },
         log: "silent",
@@ -89,6 +92,9 @@ describe("settings api", () => {
             options: {
               extended: true,
             },
+          },
+          https: {
+            enabled: false,
           },
         },
         variantHandlers: {