+++ packages/admin-api-client/CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### Breaking change
 
+## [6.2.0] - 2022-08-11
+
+### Added
+- feat(#390): Add `https` and `agent` options to the `configClient` method
+
 ## [6.1.0] - 2022-08-04
 
 ### Added