+++ packages/admin-api-paths/CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [4.2.0] - 2022-08-11
+
+### Added
+
+- feat: Add HTTPS_PROTOCOL constant
+
 
 ## [4.1.0] - 2022-08-04