+++ packages/cypress-commands/CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [6.1.0] - 2022-08-11
+
+### Added
+- feat(#390): Add `https` option to the `mocksConfigClient` command and the `AdminApiClient` class.
+- feat(#390): Add `MOCKS_SERVER_ADMIN_API_HTTPS` environment variable, allowing to enable https in the default client.
+
 ## [6.0.0] - 2022-08-04
 
 ### Added